fix: Corrected defect where decorators that depend on environment variables did not work when placed in easey-common
